% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_kodachrome.R
\name{patina_kodachrome}
\alias{patina_kodachrome}
\title{Faded, overexposed Kodachrome vibe (warm cast + halation + grain)}
\usage{
patina_kodachrome(img, warmth = 0.25, fade = 0.25, bloom = 0.35, grain = 0.6)
}
\arguments{
\item{img}{magick image.}

\item{warmth}{0..1 warm colorize.}

\item{fade}{0..1 highlight lift.}

\item{bloom}{0..1 halation glow strength.}

\item{grain}{0..1 grain amount (qualitative).}
}
\description{
Faded, overexposed Kodachrome vibe (warm cast + halation + grain)
}
