% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_journal}
\alias{scale_fill_journal}
\title{Discrete fill palette for journal themes}
\usage{
scale_fill_journal(
  era = c("1900s", "1930s", "1960s"),
  variant = c("gray", "spot"),
  ...
)
}
\arguments{
\item{era}{One of \code{"1900s"}, \code{"1930s"}, \code{"1960s"}.}

\item{variant}{\code{"gray"} (default) or \code{"spot"} (only for 1960s).}

\item{...}{Passed to \code{ggplot2::scale_color_manual()}.}
}
\value{
A discrete fill scale.
}
\description{
See \code{\link{scale_color_journal}} for details.
}
\examples{
# p + scale_fill_journal("1900s")
}
