% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_journal_continuous}
\alias{scale_fill_journal_continuous}
\title{Continuous grayscale scale (fill)}
\usage{
scale_fill_journal_continuous(low = "#0f0f0f", high = "#adadad", ...)
}
\arguments{
\item{low, high}{Hex colors for the gradient endpoints.}

\item{...}{Passed to \code{ggplot2::scale_color_gradient()}.}
}
\value{
A continuous fill scale.
}
\description{
Continuous grayscale scale (fill)
}
\examples{
\dontrun{
# p + scale_fill_journal_continuous()
}
}
