% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_journal_scan.R
\name{scanify_journal}
\alias{scanify_journal}
\title{Old journal scan preset (sepia/mono, halftone, paper texture, scanner shadow)}
\usage{
scanify_journal(img, paper = NULL, sepia = TRUE, dither = TRUE, tilt_deg = 0.6)
}
\arguments{
\item{img}{magick image}

\item{paper}{optional path to paper texture}

\item{sepia}{logical}

\item{dither}{logical}

\item{tilt_deg}{small rotation}
}
\description{
Old journal scan preset (sepia/mono, halftone, paper texture, scanner shadow)
}
