% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_edu_film}
\alias{theme_edu_film}
\title{Educational-film theme (warm page, dotted majors, gentle border)}
\usage{
theme_edu_film(
  base_size = 11,
  base_family = NULL,
  bg = "#e9e1c9",
  page = "#e5dcc5",
  grid_col = "#5b533c66"
)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{font family (optional)}

\item{bg}{panel background}

\item{page}{overall background}

\item{grid_col}{major grid color}
}
\description{
Educational-film theme (warm page, dotted majors, gentle border)
}
