% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_journal_1930s}
\alias{theme_journal_1930s}
\title{Journal theme: c. 1930s–1940s}
\usage{
theme_journal_1930s(base_size = 11, base_family = NULL)
}
\arguments{
\item{base_size}{Numeric base text size passed to \code{ggplot2::theme_minimal()}.
Defaults to 11. This does not change sizes you have already set explicitly.}

\item{base_family}{Optional base font family. Leave \code{NULL} to inherit.}
}
\value{
A \code{ggplot2::theme}.
}
\description{
Finer lines, Y and light X major grids, bordered facet strips.
}
