% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_newscast}
\alias{theme_newscast}
\title{Newscast/CRT theme (dark UI, light ink, faint grid)}
\usage{
theme_newscast(
  base_size = 11,
  base_family = NULL,
  bg = "#151515",
  page = "#111111",
  grid_col = "#ffffff22"
)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{font family (optional)}

\item{bg}{panel background}

\item{page}{overall background}

\item{grid_col}{grid color (low alpha)}
}
\description{
Newscast/CRT theme (dark UI, light ink, faint grid)
}
