% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_transparency}
\alias{theme_transparency}
\title{Transparency-slide theme (clean white, thin grid, generous margins)}
\usage{
theme_transparency(
  base_size = 11,
  base_family = NULL,
  page = "white",
  grid_col = "#bdbdbd66"
)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{font family (optional)}

\item{page}{page background}

\item{grid_col}{grid color}
}
\description{
Transparency-slide theme (clean white, thin grid, generous margins)
}
