% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpatina-package.R
\docType{package}
\name{ggpatina-package}
\alias{ggpatina}
\alias{ggpatina-package}
\title{ggpatina: Period Aesthetics and Analog Aging Effects for Figures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides period-inspired fonts and analog-aging visual effects for figures. Includes presets for transparency slides and old journal scans, plus ink bleed, Kodachrome-style fade, photocopy misalignment, and a blueprint (cyanotype) style.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jprybylski.github.io/ggpatina/}
  \item \url{https://github.com/jprybylski/ggpatina}
}

}
\author{
\strong{Maintainer}: John Prybylski \email{jprybylski@gmail.com} (\href{https://orcid.org/0000-0001-5802-0539}{ORCID}) [copyright holder]

}
\keyword{internal}
