% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_photocopy.R
\name{patina_photocopy}
\alias{patina_photocopy}
\title{Photocopier double-pass misalignment (ghost + banding)}
\usage{
patina_photocopy(
  img,
  offset = c(3, -2),
  ghost_opacity = 0.35,
  banding = 0.12,
  tilt_deg = 0.4
)
}
\arguments{
\item{img}{magick image.}

\item{offset}{length-2 integer vector (x,y) ghost offset in pixels.}

\item{ghost_opacity}{0..1 opacity of ghost pass.}

\item{banding}{0..1 banding intensity.}

\item{tilt_deg}{small rotation in degrees to mimic deskew.}
}
\description{
Photocopier double-pass misalignment (ghost + banding)
}
