% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_transparency.R
\name{slideify_transparency}
\alias{slideify_transparency}
\title{Transparency slide preset (chromatic offset, light leak, vignette, grain)}
\usage{
slideify_transparency(
  img,
  leak_strength = 0.25,
  vignette = 0.35,
  skew = 0.015,
  grain = 0.8
)
}
\arguments{
\item{img}{magick image}

\item{leak_strength}{0..1}

\item{vignette}{0..1}

\item{skew}{Keystone factor (0..~0.03)}

\item{grain}{0..1}
}
\description{
Transparency slide preset (chromatic offset, light leak, vignette, grain)
}
